+++ app/domain_models/login_user.rb
@@ -1,3 +1,4 @@
+
 class LoginUser
 
   attr_reader :uuid, :account_number, :user, :is_login
@@ -29,19 +30,4 @@ class LoginUser
     @is_login = true
   end
 
-end
-
-# todo: It would be better to arrange in another place.
-class NoUserError < StandardError
-  def initialize(account_number)
-    msg = "There is no #{account_number} user"
-    super(msg)
-  end
-end
-
-class WrongPasswordError < StandardError
-  def initialize
-    msg = "Password is wrong"
-    super(msg)
-  end
 end
